Added manual "User Themes" gnome extension install
Added some user prompts to install the "User Themes" gnome extension,
because the theming is currently buggy (likely just because this
required extension is missing.  It opens the browser to the correct
page, and tells the user what version to install.  Also added a TODO
to automate it, with a link to a page with instructions